Rename jest directory to spec
