+++ apps/browser/src/platform/popup/browser-popup-utils.spec.ts
@@ -1,4 +1,4 @@
-import { createChromeTabMock } from "../../autofill/jest/autofill-mocks";
+import { createChromeTabMock } from "../../autofill/spec/autofill-mocks";
 import { BrowserApi } from "../browser/browser-api";
 
 import BrowserPopupUtils from "./browser-popup-utils";
